fix(packages/eg-walker): validate decoder invariants and deduplicate test helpers
